+++ CHANGELOG.md
@@ -7,15 +7,34 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## Added
+
+- `documentation.tex`
+- `Floodpipe.drawio`
+- `UProperties.pas`
+- `UFunctions.pas`
+- `UTypedefine.pas`
+- `UPixelFuncitons.pas`
+
+## Modified
+
+- `UMain.pas` First layout and structure
+- `.gitignore` nomore history from delphi
+
 ## [0.1.1]
+
 ### Added
+
 - `CHANGELOG.md`
 
-### Changed
+### Modified
+
 - `README.md`
 
 ## [0.1.0] - 2022-07.22
+
 ### Added
+
 - Everything!
 
 [Unreleased]: https://github.com/thezitrox/floodpipe/compare/v0.1.1...HEAD